Move components to their own folders
