+++ src/components/Icon/index.js
@@ -1,0 +1,1 @@
+export { default as SocialIconGroup } from './SocialIconGroup';
